% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xegaPermGene.R
\name{Decay}
\alias{Decay}
\title{Exponential decay.}
\usage{
Decay(t, lambda = 0.05)
}
\arguments{
\item{t}{Number of objects.}

\item{lambda}{Exponential decay constant.}
}
\value{
Vector with t elements with values of exponential decay.
}
\description{
Exponential decay.
}
\examples{
Decay(5, 0.4)
Decay(10, 0.4)

}
\seealso{
Other Utility: 
\code{\link{without}()}
}
\concept{Utility}
