% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xegaPermlF.R
\docType{data}
\name{lFxegaPermGene}
\alias{lFxegaPermGene}
\title{Generate local functions and objects.}
\format{
An object of class \code{list} of length 21.
}
\usage{
lFxegaPermGene
}
\description{
\code{lFxegaPermGene} is a list of functions 
which contains a definition of all local objects 
required for the use of genetic operators with the 
We refer to this object as local configuration.
}
\details{
We use the local function list (the local configuration) for 
   \enumerate{
   \item
      replacing all constants with constant functions.
      
      Rationale: We need one formal argument (the local function list lF)
      and we can dispatch multiple functions. E.g.  \code{lF$verbose()}
  \item    
      for dynamically binding a local function with a definition from a
      proper function factory. E.g. the selection methods 
      \code{lf$SelectGene} and \code{SelectMate}.
      
 \item for gene representation specific special functions:
       \code{lf$InitGene}, \code{lF$DecodeGene}, \code{lf$EvalGene}
       \code{lf$ReplicateGene}, ...
      
 }
}
\seealso{
Other Configuration: 
\code{\link{xegaPermCrossoverFactory}()},
\code{\link{xegaPermMutationFactory}()}
}
\concept{Configuration}
\keyword{datasets}
