% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xegaPermDecode.R
\name{xegaPermDecodeGene}
\alias{xegaPermDecodeGene}
\title{Decode a permutation.}
\usage{
xegaPermDecodeGene(gene, lF)
}
\arguments{
\item{gene}{Permutation.}

\item{lF}{Local configuration of the genetic algorithm.}
}
\value{
A permutation gene.
}
\description{
\code{xegaPermDecodeGene} decodes a permutation gene.
}
\details{
\code{xegaPermDecodeGene} is the identy function.
}
\examples{
g<-xegaPermInitGene(lFxegaPermGene)
xegaPermDecodeGene(g)
}
\concept{Decoder}
