% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xegaPermInitGene.R
\name{xegaPermInitGene}
\alias{xegaPermInitGene}
\title{Initialize a gene with a permutation of integers}
\usage{
xegaPermInitGene(lF)
}
\arguments{
\item{lF}{Local configuration of the genetic algorithm.}
}
\value{
A permutation gene.
}
\description{
\code{xegaPermInitGene} generates a random permutation 
             with a given length n.
}
\details{
In the permutation representation of 
         package \code{xegaPerm}, a \emph{gene} is a list with 
         \enumerate{
         \item \code{$evaluated}: Boolean: TRUE if the fitness is known.
         \item \code{$fit}:       The fitness of the genotype of 
                                 \code{$gene1}.         
         \item \code{$gene1}:     The permutation (the genotype).
         }
}
\examples{
xegaPermInitGene(lFxegaPermGene)

}
\concept{Initialization}
